python/requirements.txt
pandas==2.3.3
scikit-learn==1.7.2
joblib==1.5.2
numpy==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
scipy==1.16.2
threadpoolctl==3.6.0
pyserial==3.5